+++ example/lib/main.dart
@@ -355,6 +355,23 @@ class _MyHomePageState extends State<MyHomePage> {
             ),
             ListTile(
               onTap: () {
+                Set<SimpleItem> set = Set<SimpleItem>()
+                  ..add(SimpleItem(1, "One"))
+                  ..add(SimpleItem(2, "Two"))
+                  ..add(SimpleItem(3, "Three"));
+                radioListDialog(
+                  context,
+                  "Select one",
+                  set,
+                  (item) {
+                    print(item);
+                  },
+                );
+              },
+              title: Text("Radio list dialog"),
+            ),
+            ListTile(
+              onTap: () {
                 successToast("Success toast");
               },
               title: Text("Success toast"),