+++ lib/commons.dart
@@ -10,6 +10,7 @@ export 'package:commons/src/dialogs/alert_dialogs.dart';
 export 'package:commons/src/dialogs/input_dialogs.dart';
 export 'package:commons/src/dialogs/options_dialog.dart';
 export 'package:commons/src/dialogs/selection_dialogs.dart';
+export 'package:commons/src/dialogs/radio_list_dialogs.dart';
 export 'package:commons/src/screens/stack_trace_screen.dart';
 export 'package:commons/src/functions/functions.dart';
 export 'package:commons/src/functions/extension_functions.dart';