+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.5.3] - Radio list Dialog
+
+* Added: Radio list dialog
+
 ## [0.5.2] - Selection Dialogs
 
 * Added: Single select dialog